Find directories of a family

Find directories of a family

find_family(parent, family = getOption("family") %||% "^[.]family")

parent(family = getOption("family") %||% "^[.]family")

children(family = getOption("family") %||% "^[.]family")

siblings(family = getOption("family") %||% "^[.]family", self = FALSE)

Arguments

parent
A string giving a single path to the parent directory. The children directories should be nested one level under it.

family
A regular expression to match the name of the file that defines the family.

self
Include the working directory in the output? Yes: TRUE; No: FALSE.

Value

A character vector.

Examples

library(fs)
library(family)

wd <- getwd()
on.exit(setwd(wd))

parent <- path(tempdir(), "parent")

me <- dir_create(path(parent, "me"))
sister <- dir_create(path(parent, "sister"))
brother <- dir_create(path(parent, "brother"))

# To define the family, say 'smith', add a file '.smith' to each child directory
file_create(path(me, ".smith"))
file_create(path(sister, ".smith"))
file_create(path(brother, ".smith"))

# Other directories will be ignored
neighbour <- dir_create(path(parent, "neighbour"))

dir_tree(parent)

#> /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpriDnLJ/parent
#> ├── brother
#> ├── me
#> ├── neighbour
#> └── sister

# From anywhere
family::find_family(parent, family = "^[.]smith$")

#> [1] "/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpriDnLJ/parent/brother"
#> [2] "/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpriDnLJ/parent/me"     
#> [3] "/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpriDnLJ/parent/sister" 

# You may use convenient helpers form the parent or a child:

# From the parent
setwd(parent)
family::children("^[.]smith$")

#> [1] "/private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpriDnLJ/parent/brother"
#> [2] "/private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpriDnLJ/parent/me"     
#> [3] "/private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpriDnLJ/parent/sister" 

# You may also pass `family` via `options()`
op <- options(family = "^[.]smith$")
on.exit(op, add = TRUE)

family::children()

#> [1] "/private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpriDnLJ/parent/brother"
#> [2] "/private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpriDnLJ/parent/me"     
#> [3] "/private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpriDnLJ/parent/sister" 

# From a child
setwd(me)
family::siblings()

#> [1] "/private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpriDnLJ/parent/brother"
#> [2] "/private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpriDnLJ/parent/sister" 

family::siblings(self = TRUE)

#> [1] "/private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpriDnLJ/parent/brother"
#> [2] "/private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpriDnLJ/parent/me"     
#> [3] "/private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpriDnLJ/parent/sister" 

family::parent()

#> [1] "/private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpriDnLJ/parent"

setwd(wd)